refactor(101737): :recycle: separa as regras de negocio do layout
